requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
requests>=2.31.0
pydantic>=2.4.2
pyyaml>=6.0.1
cryptography>=41.0.4
python-dotenv>=1.0.0
